CRUDs

+++ michito/src/app/Services/medicamento.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { Observable } from 'rxjs';
+import { BehaviorSubject, Observable, of } from 'rxjs';
 import { catchError } from 'rxjs/operators';
 import { Medicamento } from '../Model/medicamento';
 
@@ -12,6 +12,18 @@ export class MedicamentoService {
 
   constructor(private http: HttpClient) {}
 
+  private medicamentoSeleccionado = new BehaviorSubject<Medicamento | null>(null);
+
+  // Método para obtener el BehaviorSubject como Observable
+  obtenerMedicamentoSeleccionado(): Observable<Medicamento | null> {
+    return this.medicamentoSeleccionado.asObservable();
+  }
+
+  // Método para setear el medicamento seleccionado
+  setMedicamentoSeleccionado(medicamento: Medicamento): void {
+    this.medicamentoSeleccionado.next(medicamento);
+  }
+
   obtenerMedicamentos(): Observable<Medicamento[]> {
     console.log('Obteniendo medicamentos...');
     return this.http.get<Medicamento[]>(`${this.ROOT_URL}/all`).pipe(
@@ -21,15 +33,22 @@ export class MedicamentoService {
       })
     );
   }
+  
 
   eliminarMedicamento(id: number): Observable<any> {
-    return this.http.delete(`${this.ROOT_URL}/delete/${id}`).pipe(
-      catchError(error => {
-        console.error('Error al eliminar medicamento:', error);
-        throw error;
-      })
-    );
+    const confirmed = confirm('¿Estás seguro de que deseas eliminar este medicamento?');
+    if (confirmed) {
+      return this.http.delete(`${this.ROOT_URL}/delete/${id}`).pipe(
+        catchError(error => {
+          console.error('Error al eliminar medicamento:', error);
+          throw error;
+        })
+      );
+    } else {
+      return of(null);  // Retornar un Observable vacío si la acción es cancelada
+    }
   }
+  
 
   editarMedicamento(medicamento: Medicamento): Observable<Medicamento> {
     return this.http.put<Medicamento>(`${this.ROOT_URL}/update/${medicamento.id}`, medicamento).pipe(
@@ -39,4 +58,5 @@ export class MedicamentoService {
       })
     );
   }
+  
 }
